:recycle: Make some more seeds non-destructive.
Rename legacy seeds to travel authorization seeds.

Note that these seeds are using IDs so are very fragile.
They should be updated to load the data they need from the database.
